refactor: organize related code tests in folders
