+++ src/test/kotlin/com/atauchi/transformerFileService/core/application/transformerFile/TransformerFileServiceTests.kt
@@ -1,7 +1,5 @@
-package com.atauchi.transformerFileService.core.application
+package com.atauchi.transformerFileService.core.application.transformerFile
 
-import com.atauchi.transformerFileService.core.application.transformerFile.TransformerFileService
-import com.atauchi.transformerFileService.core.application.transformerFile.TxtFileParser
 import com.atauchi.transformerFileService.core.domain.entities.transformerFile.File
 import com.atauchi.transformerFileService.core.domain.entities.transformerFile.User
 import com.atauchi.transformerFileService.infra.db.mongo.FileDocument
