+++ src/test/kotlin/com/atauchi/transformerFileService/core/application/transformerFile/TxtFileParserTests.kt
@@ -1,6 +1,5 @@
-package com.atauchi.transformerFileService.core.application
+package com.atauchi.transformerFileService.core.application.transformerFile
 
-import com.atauchi.transformerFileService.core.application.transformerFile.TxtFileParser
 import com.atauchi.transformerFileService.core.domain.entities.transformerFile.Order
 import com.atauchi.transformerFileService.core.domain.entities.transformerFile.Product
 import com.atauchi.transformerFileService.core.domain.entities.transformerFile.User
